+++ cluster_tunnel/config.py
@@ -58,6 +58,7 @@ class Cluster(BaseModel):
     ssh_alias: Optional[str] = None
     identity_file: Optional[str] = None
     requires_otp: bool = False
+    requires_password: bool = True
     control_persist: Optional[str] = None
     server_alive_interval: Optional[int] = None
     server_alive_count_max: Optional[int] = None